log: add a patch to not exit azurecli if stdout is not drained
