+++ lib/cli.js
@@ -163,8 +163,12 @@ _.extend(AzureCli.prototype, {
     if (self.uncaughtExceptionHandler) {
       process.removeListener('uncaughtException', self.uncaughExceptionHandler);
     }
-
-    process.exit(exitCode);
+    if (log.shouldWaitForStdoutDrained) {
+      process.on('exit', function () { process.exit(exitCode); });
+    }
+    else {
+      process.exit(exitCode);
+    }
   },
   
   normalizeAuthorizationError: function (msg) {
